refactor(card): extract evolution collector
